+++ test/shopify_api/scopes_test.exs
@@ -1,0 +1,5 @@
+defmodule ShopifyAPI.ScopesTest do
+  use ExUnit.Case, async: true
+
+  doctest ShopifyAPI.Scopes
+end